NN-2325 Add validation
