+++ backend/controllers/bulkAppointmentsAddDetails.js
@@ -4,7 +4,7 @@ const { serviceUnavailableMessage } = require('../common-messages')
 const { DATE_TIME_FORMAT_SPEC, DAY_MONTH_YEAR, buildDateTime } = require('../../src/dateHelpers')
 const { capitalize } = require('../utils')
 const { calculateEndDate } = require('../../src/BulkAppointments/RecurringAppointments')
-const { repeatTypes } = require('../shared/appointmentConstants')
+const { repeatTypes, getValidationMessages } = require('../shared/appointmentConstants')
 
 const {
   app: { notmEndpointUrl },
@@ -47,87 +47,6 @@ const endRecurringEndingDate = ({ date, startTime, times, repeats }) => {
     endOfPeriod,
     recurringStartTime,
   }
-}
-
-const getValidationMessages = ({
-  appointmentType,
-  location,
-  date,
-  startTime,
-  endTime,
-  sameTimeAppointments,
-  comments,
-  recurring,
-  repeats,
-  times,
-}) => {
-  const errors = []
-  const now = moment()
-  const isToday = date ? moment(date, DAY_MONTH_YEAR).isSame(now, 'day') : false
-
-  if (!appointmentType) errors.push({ text: 'Select an appointment type', href: '#appointment-type' })
-
-  if (!location) errors.push({ text: 'Select a location', href: '#location' })
-
-  if (!date) errors.push({ text: 'Select a date', href: '#date' })
-
-  if (!sameTimeAppointments)
-    errors.push({ text: 'Select yes if the appointments all have the same time', href: '#same-time-appointments' })
-
-  if (date && !moment(date, DAY_MONTH_YEAR).isValid())
-    errors.push({ text: 'Enter a date in DD/MM/YYYY format', href: '#date' })
-
-  if (date && moment(date, DAY_MONTH_YEAR).isBefore(now, 'day'))
-    errors.push({ text: 'Select a date that is not in the past', href: '#date' })
-
-  if (sameTimeAppointments === 'yes') {
-    const startTimeDuration = moment.duration(now.diff(startTime))
-    const endTimeDuration = endTime && moment.duration(startTime.diff(endTime))
-
-    if (!startTime) errors.push({ text: 'Select a start time', href: '#start-time-hours' })
-
-    if (isToday && startTimeDuration.asMinutes() > 1)
-      errors.push({ text: 'Select a start time that is not in the past', href: '#start-time-hours' })
-
-    if (endTime && endTimeDuration.asMinutes() > 1) {
-      errors.push({ text: 'Select an end time that is not in the past', href: '#end-time-hours' })
-    }
-  }
-
-  if (comments && comments.length > 3600)
-    errors.push({ text: 'Maximum length should not exceed 3600 characters', href: '#comments' })
-
-  if (!recurring) errors.push({ href: '#recurring', text: 'Select yes if these are recurring appointments' })
-  if (recurring === 'yes' && !repeats) errors.push({ href: '#repeats', text: 'Select a period' })
-
-  if (recurring === 'yes') {
-    if (Number(times) <= 0 || !Number(times))
-      errors.push({ href: '#times', text: 'Enter the number of appointments using numbers only' })
-
-    if (repeats && times) {
-      const { recurringStartTime, endOfPeriod } = endRecurringEndingDate({ date, startTime, repeats, times })
-
-      if (endOfPeriod && endOfPeriod.isSameOrAfter(recurringStartTime.startOf('day').add(1, 'years'))) {
-        errors.push({
-          href: '#times',
-          text: 'Select fewer number of appointments - you can only add them for a maximum of 1 year',
-        })
-      }
-    }
-
-    if (repeats === 'WEEKDAYS') {
-      const SATURDAY = 6
-      const SUNDAY = 0
-      if (moment(date, DAY_MONTH_YEAR).day() === SATURDAY || moment(date, DAY_MONTH_YEAR).day() === SUNDAY) {
-        errors.push({
-          href: '#date',
-          text: 'The date must be a week day',
-        })
-      }
-    }
-  }
-
-  return errors
 }
 
 const bulkAppointmentsAddDetailsFactory = (appointmentsService, oauthApi, logError) => {
@@ -218,18 +137,23 @@ const bulkAppointmentsAddDetailsFactory = (appointmentsService, oauthApi, logErr
       const appointmentTypeDetails = appointmentType && appointmentTypes.find(type => type.value === appointmentType)
       const locationDetails = location && locations.find(type => type.value === Number(location))
 
-      const errors = getValidationMessages({
-        appointmentType,
-        location,
-        date,
-        startTime,
-        endTime,
-        sameTimeAppointments,
-        comments,
-        times,
-        repeats,
-        recurring,
-      })
+      const errors = [
+        ...getValidationMessages({
+          appointmentType,
+          location,
+          date,
+          startTime,
+          endTime,
+          sameTimeAppointments,
+          comments,
+          times,
+          repeats,
+          recurring,
+        }),
+      ]
+
+      if (!sameTimeAppointments)
+        errors.push({ text: 'Select yes if the appointments all have the same time', href: '#same-time-appointments' })
 
       const { endOfPeriod } = endRecurringEndingDate({ date, startTime, repeats, times })
