+++ views/addAppointment.njk
@@ -19,14 +19,14 @@
           label: {
               text: "Repeats"
           },
-          items: repeatTypes | addDefaultSelectedVale('Select') | setSelected(repeats),
+          items: repeatTypes | addDefaultSelectedVale('Select') | setSelected(formValues.repeats),
           errorMessage: errors | findError('repeats')
       }) }}
 
       {{ govukInput({
           id: "times",
           name: "times",
-          value: times,
+          value: formValues.times,
           errorMessage: errors | findError('times'),
           type: "number",
           label: {
@@ -85,30 +85,30 @@
           <input type="hidden" name="bookingId" value={{bookingId}} />
 
           {{ govukSelect({
-            id: "appointmentType",
+            id: "appointment-type",
             name: "appointmentType",
             label: {
                 text: "Appointment type"
             },
-            items: appointmentTypes | addDefaultSelectedVale('Select'),
+            items: appointmentTypes | addDefaultSelectedVale('Select') | setSelected(formValues.appointmentType),
             errorMessage: errors | findError('appointment-type')
           }) }}
 
           {{ govukSelect({
-            id: "appointmentLocation",
-            name: "appointmentLocation",
-            label: {
-                text: "Appointment location"
-            },
-            items: appointmentLocations | addDefaultSelectedVale('Select'),
-            errorMessage: errors | findError('location')
+              id: "location",
+              name: "location",
+              label: {
+                  text: "Location"
+              },
+              items: appointmentLocations | addDefaultSelectedVale('Select') | setSelected(formValues.location),
+              errorMessage: errors | findError('location')
           }) }}
 
           {{ datePicker({
               id: 'date',
               label: 'Date',
               name: 'date',
-              date: date,
+              date: formValues.date,
               errorMessage: errors | findError('date'),
               attributes: {'data-disable-past-date':'true'}
           }) }}
@@ -117,8 +117,8 @@
               id: 'start-time',
               label: 'Start time',
               name: 'startTime',
-              hour: startTimeHours,
-              minute: startTimeMinutes,
+              hour: formValues.startTimeHours,
+              minute: formValues.startTimeMinutes,
               errorMessage: errors | findError('start-time-hours')
           }) }}
 
@@ -126,8 +126,8 @@
               id: 'end-time',
               label: 'End time (optional)',
               name: 'endTime',
-              hour: endTimeHours,
-              minute: endTimeMinutes,
+              hour: formValues.endTimeHours,
+              minute: formValues.endTimeMinutes,
               errorMessage: errors | findError('end-time-hours')
           }) }}
 
@@ -145,7 +145,7 @@
               items: [{
                   value: "yes",
                   text: "Yes",
-                  checked: recurring == 'yes',
+                  checked: formValues.recurring == 'yes',
                   conditional: {
                       html: recurringHtml
                   }
@@ -153,7 +153,7 @@
                   {
                       value: "no",
                       text: "No",
-                      checked: recurring == 'no'
+                      checked: formValues.recurring == 'no'
                   }
               ]
           }) }}
@@ -161,7 +161,7 @@
           {{ govukTextarea({
               name: "comments",
               id: "comments",
-              value: comments,
+              value: formValues.comments,
               errorMessage: errors | findError('comments'),
               label: {
                   text: "Comments (optional)"
